requirements.txt
Flask
python-dotenv
requests
openai
gunicorn
Flask-Corss